requirements.txt
pandas
joblib
streamlit
Pillow
